refactor: Mock 객체를 Unit Test 타켓에 위치하도록 변경
