+++ OpenMarket/OpenMarket.xcodeproj/xcshareddata/xcschemes/OpenMarket.xcscheme
@@ -20,6 +20,20 @@
                ReferencedContainer = "container:OpenMarket.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
+         <BuildActionEntry
+            buildForTesting = "NO"
+            buildForRunning = "NO"
+            buildForProfiling = "NO"
+            buildForArchiving = "NO"
+            buildForAnalyzing = "NO">
+            <BuildableReference
+               BuildableIdentifier = "primary"
+               BlueprintIdentifier = "205D32C0287D6084004CF76F"
+               BuildableName = "MockURLSessionTests.xctest"
+               BlueprintName = "MockURLSessionTests"
+               ReferencedContainer = "container:OpenMarket.xcodeproj">
+            </BuildableReference>
+         </BuildActionEntry>
       </BuildActionEntries>
    </BuildAction>
    <TestAction
@@ -28,16 +42,6 @@
       selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
       shouldUseLaunchSchemeArgsEnv = "YES">
       <Testables>
-         <TestableReference
-            skipped = "NO">
-            <BuildableReference
-               BuildableIdentifier = "primary"
-               BlueprintIdentifier = "205D32C0287D6084004CF76F"
-               BuildableName = "MockURLSessionTests.xctest"
-               BlueprintName = "MockURLSessionTests"
-               ReferencedContainer = "container:OpenMarket.xcodeproj">
-            </BuildableReference>
-         </TestableReference>
          <TestableReference
             skipped = "NO">
             <BuildableReference
